+++ src/main/java/net/spell_engine/internals/SpellHelper.java
@@ -22,6 +22,7 @@ import net.spell_engine.api.spell.CustomSpellHandler;
 import net.spell_engine.api.spell.Spell;
 import net.spell_engine.api.spell.SpellEvents;
 import net.spell_engine.api.spell.SpellInfo;
+import net.spell_engine.compat.QuiverCompat;
 import net.spell_engine.entity.ConfigurableKnockback;
 import net.spell_engine.entity.SpellCloud;
 import net.spell_engine.entity.SpellProjectile;
@@ -82,6 +83,9 @@ public class SpellHelper {
             if(ammoItem != null) {
                 ammo = ammoItem.getDefaultStack();
                 satisfied = player.getInventory().contains(ammo);
+                if (needsArrow) {
+                    satisfied = satisfied || QuiverCompat.hasArrow(ammoItem, player);
+                }
             }
         }
         return new AmmoResult(satisfied, ammo);